Fixed typo in help
Note: $PSBoundParameters and corresponding variables in the Begin
section are not available if the function arguments were passed via
pipeline. So if one ever happen to implement support for the
ValueFromPipeline attribute, the code that recreates variables from the
bound parameters in examples will need to be moved to the Process block.